+++ api/cmd/worker/main.go
@@ -3,6 +3,9 @@ package main
 import (
 	"context"
 	"log"
+	"os"
+	"os/signal"
+	"syscall"
 
 	"github.com/jtonynet/go-scheduler-trigger/api/bootstrap"
 	"github.com/jtonynet/go-scheduler-trigger/api/config"
@@ -10,7 +13,12 @@ import (
 )
 
 func main() {
-	ctx := context.Background()
+	ctx, stop := signal.NotifyContext(
+		context.Background(),
+		os.Interrupt,
+		syscall.SIGTERM,
+	)
+	defer stop()
 
 	cfg, err := config.LoadConfig(".")
 	if err != nil {