+++ api/config/config.go
@@ -1,6 +1,8 @@
 package config
 
 import (
+	"time"
+
 	"github.com/spf13/viper"
 )
 
@@ -23,13 +25,21 @@ type PubSub struct {
 }
 
 type InMemoryDatabase struct {
-	Strategy   string
-	Pass       string
-	Port       string
-	Host       string
-	DB         int
-	Protocol   int
-	Expiration int
+	Strategy        string
+	Pass            string
+	Port            string
+	Host            string
+	DB              int
+	Protocol        int
+	Expiration      time.Duration
+	MaxRetries      int
+	MinRetryBackoff time.Duration
+	MaxRetryBackoff time.Duration
+	DialTimeout     time.Duration
+	ReadTimeout     time.Duration
+	WriteTimeout    time.Duration
+	PoolSize        int
+	MinIdleConns    int
 }
 
 type InMemoryDatabaseConverter interface {
@@ -37,46 +47,78 @@ type InMemoryDatabaseConverter interface {
 }
 
 type Trigger struct {
-	Strategy   string `mapstructure:"TRIGGER_IN_MEMORY_STRATEGY"`
-	Pass       string `mapstructure:"TRIGGER_IN_MEMORY_PASSWORD"`
-	Port       string `mapstructure:"TRIGGER_IN_MEMORY_PORT"`
-	Host       string `mapstructure:"TRIGGER_IN_MEMORY_HOST"`
-	DB         int    `mapstructure:"TRIGGER_IN_MEMORY_DB"`
-	Protocol   int    `mapstructure:"TRIGGER_IN_MEMORY_PROTOCOL"`
-	Expiration int    `mapstructure:"TRIGGER_IN_MEMORY_EXPIRATION_DEFAULT_IN_MS"`
+	Strategy        string        `mapstructure:"TRIGGER_IN_MEMORY_STRATEGY"`
+	Pass            string        `mapstructure:"TRIGGER_IN_MEMORY_PASSWORD"`
+	Port            string        `mapstructure:"TRIGGER_IN_MEMORY_PORT"`
+	Host            string        `mapstructure:"TRIGGER_IN_MEMORY_HOST"`
+	DB              int           `mapstructure:"TRIGGER_IN_MEMORY_DB"`
+	Protocol        int           `mapstructure:"TRIGGER_IN_MEMORY_PROTOCOL"`
+	Expiration      time.Duration `mapstructure:"TRIGGER_IN_MEMORY_EXPIRATION_DEFAULT_IN_MS"`
+	MaxRetries      int           `mapstructure:"TRIGGER_MAX_RETRIES"`
+	MinRetryBackoff time.Duration `mapstructure:"TRIGGER_MIN_RETRY_BACKOFF_IN_MS"`
+	MaxRetryBackoff time.Duration `mapstructure:"TRIGGER_MAX_RETRY_BACKOFF_IN_S"`
+	DialTimeout     time.Duration `mapstructure:"TRIGGER_DIAL_TIMEOUT_IN_S"`
+	ReadTimeout     time.Duration `mapstructure:"TRIGGER_READ_TIMEOUT_IN_S"`
+	WriteTimeout    time.Duration `mapstructure:"TRIGGER_WRITE_TIMEOUT_IN_S"`
+	PoolSize        int           `mapstructure:"TRIGGER_POOL_SIZE"`
+	MinIdleConns    int           `mapstructure:"TRIGGER_MIN_IDDLE_CONNS"`
 }
 
 func (t *Trigger) ToInMemoryDB() InMemoryDatabase {
 	return InMemoryDatabase{
-		Strategy:   t.Strategy,
-		Pass:       t.Pass,
-		Port:       t.Port,
-		Host:       t.Host,
-		DB:         t.DB,
-		Protocol:   t.Protocol,
-		Expiration: t.Expiration,
+		Strategy:        t.Strategy,
+		Pass:            t.Pass,
+		Port:            t.Port,
+		Host:            t.Host,
+		DB:              t.DB,
+		Protocol:        t.Protocol,
+		Expiration:      t.Expiration,
+		MaxRetries:      t.MaxRetries,
+		MinRetryBackoff: t.MinRetryBackoff,
+		MaxRetryBackoff: t.MaxRetryBackoff,
+		DialTimeout:     t.DialTimeout,
+		ReadTimeout:     t.ReadTimeout,
+		WriteTimeout:    t.WriteTimeout,
+		PoolSize:        t.PoolSize,
+		MinIdleConns:    t.MinIdleConns,
 	}
 }
 
 type ShadowKey struct {
-	Strategy   string `mapstructure:"SHADOWKEY_IN_MEMORY_STRATEGY"`
-	Pass       string `mapstructure:"SHADOWKEY_IN_MEMORY_PASSWORD"`
-	Port       string `mapstructure:"SHADOWKEY_IN_MEMORY_PORT"`
-	Host       string `mapstructure:"SHADOWKEY_IN_MEMORY_HOST"`
-	DB         int    `mapstructure:"SHADOWKEY_IN_MEMORY_DB"`
-	Protocol   int    `mapstructure:"SHADOWKEY_IN_MEMORY_PROTOCOL"`
-	Expiration int    `mapstructure:"SHADOWKEY_IN_MEMORY_EXPIRATION_DEFAULT_IN_MS"`
+	Strategy        string        `mapstructure:"SHADOWKEY_IN_MEMORY_STRATEGY"`
+	Pass            string        `mapstructure:"SHADOWKEY_IN_MEMORY_PASSWORD"`
+	Port            string        `mapstructure:"SHADOWKEY_IN_MEMORY_PORT"`
+	Host            string        `mapstructure:"SHADOWKEY_IN_MEMORY_HOST"`
+	DB              int           `mapstructure:"SHADOWKEY_IN_MEMORY_DB"`
+	Protocol        int           `mapstructure:"SHADOWKEY_IN_MEMORY_PROTOCOL"`
+	Expiration      time.Duration `mapstructure:"SHADOWKEY_IN_MEMORY_EXPIRATION_DEFAULT_IN_MS"`
+	MaxRetries      int           `mapstructure:"SHADOWKEY_MAX_RETRIES"`
+	MinRetryBackoff time.Duration `mapstructure:"SHADOWKEY_MIN_RETRY_BACKOFF_IN_MS"`
+	MaxRetryBackoff time.Duration `mapstructure:"SHADOWKEY_MAX_RETRY_BACKOFF_IN_S"`
+	DialTimeout     time.Duration `mapstructure:"SHADOWKEY_DIAL_TIMEOUT_IN_S"`
+	ReadTimeout     time.Duration `mapstructure:"SHADOWKEY_READ_TIMEOUT_IN_S"`
+	WriteTimeout    time.Duration `mapstructure:"SHADOWKEY_WRITE_TIMEOUT_IN_S"`
+	PoolSize        int           `mapstructure:"SHADOWKEY_POOL_SIZE"`
+	MinIdleConns    int           `mapstructure:"SHADOWKEY_MIN_IDDLE_CONNS"`
 }
 
-func (c *ShadowKey) ToInMemoryDB() InMemoryDatabase {
+func (s *ShadowKey) ToInMemoryDB() InMemoryDatabase {
 	return InMemoryDatabase{
-		Strategy:   c.Strategy,
-		Pass:       c.Pass,
-		Port:       c.Port,
-		Host:       c.Host,
-		DB:         c.DB,
-		Protocol:   c.Protocol,
-		Expiration: c.Expiration,
+		Strategy:        s.Strategy,
+		Pass:            s.Pass,
+		Port:            s.Port,
+		Host:            s.Host,
+		DB:              s.DB,
+		Protocol:        s.Protocol,
+		Expiration:      s.Expiration,
+		MaxRetries:      s.MaxRetries,
+		MinRetryBackoff: s.MinRetryBackoff,
+		MaxRetryBackoff: s.MaxRetryBackoff,
+		DialTimeout:     s.DialTimeout,
+		ReadTimeout:     s.ReadTimeout,
+		WriteTimeout:    s.WriteTimeout,
+		PoolSize:        s.PoolSize,
+		MinIdleConns:    s.MinIdleConns,
 	}
 }