+++ CHANGELOG.md
@@ -10,6 +10,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 
+## [0.0.5] - 2025-16-12
+### Added
+
+- CMDs with Gracefull Shutdown
+- Redis Database In Memory with reconnection
+- Redis PubSub with reconnection and backoff 
+
 ## [0.0.4] - 2025-14-12
 ### Added
 
@@ -38,6 +45,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - First commit to start the project/study.
 
+[0.0.5]: https://github.com/jtonynet/go-scheduler-trigger/compare/v0.0.4...v0.0.5
 [0.0.4]: https://github.com/jtonynet/go-scheduler-trigger/compare/v0.0.3...v0.0.4
 [0.0.3]: https://github.com/jtonynet/go-scheduler-trigger/compare/v0.0.2...v0.0.3
 [0.0.2]: https://github.com/jtonynet/go-scheduler-trigger/compare/v0.0.1...v0.0.2